ShoppingCart/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
flask-marshmallow==1.2.1
marshmallow-sqlalchemy==1.0.0
requests==2.31.0
